pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
    "snscrape>=0.7.0.20230622",
    "zenrows>=1.4.0",
]
